Corrected minor bug in IM483I and some warnings in 64 bit

+++ Ball.cpp
@@ -22,7 +22,7 @@
 
 THREAD_PROC_RETURN_VALUE BallThread(void* pParam)
 {
-	int id = (intptr_t)pParam;
+	int id = (int)(intptr_t)pParam;
 
 	FILE* logballfile = NULL;
 	char logballfilename[MAX_BUF_LEN];
